+++ src/main/java/Duke.java
@@ -1,3 +1,8 @@
+import duke.command.Parser;
+import duke.io.Storage;
+import duke.task.TaskList;
+import duke.ui.Ui;
+
 import java.io.FileNotFoundException;
 import java.util.Scanner;
