+++ src/main/java/duke/task/Deadline.java
@@ -1,31 +1,35 @@
+package duke.task;
+
+import duke.task.Task;
+
 import java.io.FileWriter;
 import java.io.IOException;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 
-public class Event extends Task{
+public class Deadline extends Task {
     private String taskName;
-    private String at;
-    private String type = "E";
+    private String by;
     private LocalDate date;
     private int time;
     private int day;
     private int month;
     private int year;
-    private String[] TaskAtSplit;
+    private String type = "D";
+    private String[] TaskBySplit;
 
-    public Event(String taskName) {
-        String[] TaskAtSplit = taskName.split("/", 2);
+    public Deadline(String taskName) {
+        String[] TaskBySplit = taskName.split("/", 2);
 
         if (taskName.contains("|")) {
-            TaskAtSplit = taskName.split("\\|", 2);
-        } else if (taskName.contains("/at")) {
-            TaskAtSplit = taskName.split("/at", 2);
+            TaskBySplit = taskName.split("\\|", 2);
+        } else if (taskName.contains("/by")) {
+            TaskBySplit = taskName.split("/by", 2);
         }
 
-        this.taskName = TaskAtSplit[0].trim();
-        this.at = TaskAtSplit[1].trim();
-        String[] dateTimeSplit = this.at.split(" ", 2);
+        this.taskName = TaskBySplit[0].trim();
+        this.by = TaskBySplit[1].trim();
+        String[] dateTimeSplit = this.by.split(" ", 2);
         this.time = Integer.parseInt(dateTimeSplit[1]);
 
         if (dateTimeSplit[0].contains("/")) {
@@ -40,24 +44,28 @@ public class Event extends Task{
     }
 
     public String showTask() {
-        return this.taskName + " (at: " + this.date.format(DateTimeFormatter.ofPattern("MMM d yyyy")) + ", " + this.time + ")";
+        return this.taskName + " (by: " + this.date.format(DateTimeFormatter.ofPattern("MMM d yyyy")) + ", " + this.time + ")";
     }
 
-    public String showTaskOnly() { return this.taskName; }
+    public String showTaskOnly() {
+        return this.taskName;
+    }
 
-    public String showWhen() { return this.at; }
+    public String showWhen() {
+        return this.by;
+    }
 
     public String showType() {
         return this.type;
     }
+
+    public LocalDate showDate() { return this.date; }
+
+    public int showTime() { return this.time; }
 
     public void appendToFile(String filePath, String textToAppend) throws IOException {
         FileWriter fw = new FileWriter(filePath, true);
         fw.write(textToAppend);
         fw.close();
     }
-
-    public LocalDate showDate() { return this.date; }
-
-    public int showTime() { return this.time; }
 }
