+++ src/main/java/duke/task/Event.java
@@ -1,32 +1,33 @@
+package duke.task;
+
 import java.io.FileWriter;
 import java.io.IOException;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
-import java.time.LocalTime;
 
-public class Deadline extends Task {
+public class Event extends Task {
     private String taskName;
-    private String by;
+    private String at;
+    private String type = "E";
     private LocalDate date;
     private int time;
     private int day;
     private int month;
     private int year;
-    private String type = "D";
-    private String[] TaskBySplit;
+    private String[] TaskAtSplit;
 
-    public Deadline(String taskName) {
-        String[] TaskBySplit = taskName.split("/", 2);
+    public Event(String taskName) {
+        String[] TaskAtSplit = taskName.split("/", 2);
 
         if (taskName.contains("|")) {
-            TaskBySplit = taskName.split("\\|", 2);
-        } else if (taskName.contains("/by")) {
-            TaskBySplit = taskName.split("/by", 2);
+            TaskAtSplit = taskName.split("\\|", 2);
+        } else if (taskName.contains("/at")) {
+            TaskAtSplit = taskName.split("/at", 2);
         }
 
-        this.taskName = TaskBySplit[0].trim();
-        this.by = TaskBySplit[1].trim();
-        String[] dateTimeSplit = this.by.split(" ", 2);
+        this.taskName = TaskAtSplit[0].trim();
+        this.at = TaskAtSplit[1].trim();
+        String[] dateTimeSplit = this.at.split(" ", 2);
         this.time = Integer.parseInt(dateTimeSplit[1]);
 
         if (dateTimeSplit[0].contains("/")) {
@@ -41,28 +42,24 @@ public class Deadline extends Task {
     }
 
     public String showTask() {
-        return this.taskName + " (by: " + this.date.format(DateTimeFormatter.ofPattern("MMM d yyyy")) + ", " + this.time + ")";
+        return this.taskName + " (at: " + this.date.format(DateTimeFormatter.ofPattern("MMM d yyyy")) + ", " + this.time + ")";
     }
 
-    public String showTaskOnly() {
-        return this.taskName;
-    }
+    public String showTaskOnly() { return this.taskName; }
 
-    public String showWhen() {
-        return this.by;
-    }
+    public String showWhen() { return this.at; }
 
     public String showType() {
         return this.type;
     }
-
-    public LocalDate showDate() { return this.date; }
-
-    public int showTime() { return this.time; }
 
     public void appendToFile(String filePath, String textToAppend) throws IOException {
         FileWriter fw = new FileWriter(filePath, true);
         fw.write(textToAppend);
         fw.close();
     }
+
+    public LocalDate showDate() { return this.date; }
+
+    public int showTime() { return this.time; }
 }
